+++ server/routes/search_relevance_route_service.ts
@@ -184,6 +184,8 @@ export function registerSearchRelevanceRoutes(router: IRouter): void {
           contextFields: schema.maybe(schema.arrayOf(schema.string())),
           clickModel: schema.maybe(schema.string()),
           maxRank: schema.maybe(schema.number()),
+          startDate: schema.maybe(schema.string()),
+          endDate: schema.maybe(schema.string())
         }),
       },
     },